Update library design to allow for better composition
Currently, it's not possible to load both typescript and jsx, because
both of them extend index and make changes to the base config, but when
one is loaded after the other, the changes they made on top of index are
overridden by the index loaded by the latter.

This commit makes it so no presets explicitly load the index anymore,
leaving it to the user to decide on the load order.