Update approach to testing
Currently, we run the tests per preset, but since every module loads the
eslint base, every preset will generate a warning when base rules are
missing.

The new approach combines all configuration in a single testing preset
that contains all modules. Conflicts are not a problem, because we're
only interested in whethe rules are defined anywhere, and a conflict
will just lead to an override. I'll update #11 with the new test output.