Allow multiple JSX nodes per line
This rule, when enabled, prevents using React's string interpolation,
for example:

    <a>
      Hello {thing}!
    </a>

Would have to be changed to

    <a>
      {`Hello ${thing}!`}
    </a>

And while the rule itself is useful, its primary use is already covered
by jsx-closing-bracket-location.

This commit disables the rule to allow string interpolation, and we'll
rely on jsx-closing-bracket-location to prevent multiple opening or
closing tags on the same line.