Demote max-* rules from error to warning
The idea here is that these rules indicate that some code needs to be
refactored, but can be committed. That way, in one commit the limits can
be exceeded, and fixed in the next commit. This prevents cases where
something has to be refactored before something new is added, which the
dev only finds out while adding the new thing.