Add a package lock for reproducible builds
Right now, the burden of maintenance is increased by not having a
package lock, because every time Dependabot opens a pull request, the
build fails due to deprecated or newly added rules.

With the package lock in place, one fix can be treated in isolation from
others.

The drawback is that we'll have to proactively bump our dependencies to
find newly added or deprecated rules.